Revert "Plan 9 from Bell Labs Fourth Edition 2008-01-20"
This reverts commit adb31a6221c09bb6f8b2c527bf9100e20ffac78f.

+++ sys/src/cmd/ndb/dnserver.c
@@ -165,7 +165,7 @@ dnserver(DNSmsg *reqp, DNSmsg *repp, Request *req, uchar *srcip, int rcode)
 static RR*
 doextquery(DNSmsg *mp, Request *req, int recurse)
 {
-	ushort type;
+	int type;
 	char *name;
 	RR *rp, *neg;
